Refactor main.go into smaller logical files
- Split main.go into separate files based on responsibility:
  - config.go: Configuration and environment variables
  - fetcher.go: Template fetching from Git/Gist/GitHub
  - main.go: Entry point with minimal setup code
  - resolver.go: Core resolver implementation
  - server.go: HTTP server for standalone mode
  - template.go: Template rendering and YAML utilities
  - types.go: Resource and interface definitions
  - utils.go: Helper functions

- This improves code organization and maintainability
- No functional changes to behavior
