+++ cmd/template-resolver/resolver_test.go
@@ -3,7 +3,7 @@ package main
 import (
 	"context"
 	"testing"
-	
+
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	pipelinev1 "github.com/tektoncd/pipeline/pkg/apis/pipeline/v1"
@@ -44,12 +44,12 @@ spec:
 `,
 		},
 	}
-	
+
 	// Create resolver with mock fetcher
 	r := &resolver{
 		fetcher: mockData,
 	}
-	
+
 	// Test with basic parameters
 	params := []pipelinev1.Param{
 		{
@@ -74,14 +74,14 @@ spec:
 			},
 		},
 	}
-	
+
 	// Execute the Resolve function
 	result, err := r.Resolve(context.Background(), params)
-	
+
 	// Verify results
 	require.NoError(t, err)
 	require.NotNil(t, result)
-	
+
 	// Check that the template was rendered
 	renderedData := string(result.Data())
 	assert.Contains(t, renderedData, "name: test-pipeline")
@@ -136,12 +136,12 @@ spec:
 `,
 		},
 	}
-	
+
 	// Create resolver with mock fetcher
 	r := &resolver{
 		fetcher: mockData,
 	}
-	
+
 	// Test with custom validation steps parameter
 	params := []pipelinev1.Param{
 		{
@@ -179,14 +179,14 @@ params:
 			},
 		},
 	}
-	
+
 	// Execute the Resolve function
 	result, err := r.Resolve(context.Background(), params)
-	
+
 	// Verify results
 	require.NoError(t, err)
 	require.NotNil(t, result)
-	
+
 	// Check that the template was rendered with our custom steps
 	renderedData := string(result.Data())
 	assert.Contains(t, renderedData, "name: validation-step-1")
@@ -258,12 +258,12 @@ spec:
 `,
 		},
 	}
-	
+
 	// Create resolver with mock fetcher
 	r := &resolver{
 		fetcher: mockData,
 	}
-	
+
 	// Test with both array and string parameters containing tasks
 	params := []pipelinev1.Param{
 		{
@@ -325,14 +325,14 @@ params:
 			},
 		},
 	}
-	
+
 	// Execute the Resolve function
 	result, err := r.Resolve(context.Background(), params)
-	
+
 	// Verify results
 	require.NoError(t, err)
 	require.NotNil(t, result)
-	
+
 	// Check that the template was rendered with both task types
 	renderedData := string(result.Data())
 	assert.Contains(t, renderedData, "name: custom-validation")
@@ -408,12 +408,12 @@ spec:
 `,
 		},
 	}
-	
+
 	// Create resolver with mock fetcher
 	r := &resolver{
 		fetcher: mockData,
 	}
-	
+
 	// Test with multiple custom task parameters
 	params := []pipelinev1.Param{
 		{
@@ -459,14 +459,14 @@ params:
 			},
 		},
 	}
-	
+
 	// Execute the Resolve function
 	result, err := r.Resolve(context.Background(), params)
-	
+
 	// Verify results
 	require.NoError(t, err)
 	require.NotNil(t, result)
-	
+
 	// Check that the template was rendered with both custom step types
 	renderedData := string(result.Data())
 	assert.Contains(t, renderedData, "name: security-scan")
@@ -502,12 +502,12 @@ spec:
 `,
 		},
 	}
-	
+
 	// Create resolver with mock fetcher
 	r := &resolver{
 		fetcher: mockData,
 	}
-	
+
 	// Test with a regular array parameter
 	params := []pipelinev1.Param{
 		{
@@ -532,14 +532,14 @@ spec:
 			},
 		},
 	}
-	
+
 	// Execute the Resolve function
 	result, err := r.Resolve(context.Background(), params)
-	
+
 	// Verify results
 	require.NoError(t, err)
 	require.NotNil(t, result)
-	
+
 	// Check that the template was rendered with the array values
 	renderedData := string(result.Data())
 	assert.Contains(t, renderedData, "- dev")
@@ -632,12 +632,12 @@ spec:
 `,
 		},
 	}
-	
+
 	// Create resolver with mock fetcher
 	r := &resolver{
 		fetcher: mockData,
 	}
-	
+
 	// Test with lists of YAML objects
 	params := []pipelinev1.Param{
 		{
@@ -708,17 +708,17 @@ spec:
 			},
 		},
 	}
-	
+
 	// Execute the Resolve function
 	result, err := r.Resolve(context.Background(), params)
-	
+
 	// Verify results
 	require.NoError(t, err)
 	require.NotNil(t, result)
-	
+
 	// Check that the template was rendered with all the objects from the lists
 	renderedData := string(result.Data())
-	
+
 	// Verify environment config rendering
 	assert.Contains(t, renderedData, "name: deploy-to-development")
 	assert.Contains(t, renderedData, "value: development")
@@ -726,26 +726,26 @@ spec:
 	assert.Contains(t, renderedData, "value: app-dev")
 	assert.Contains(t, renderedData, "value: \"1\"")
 	assert.Contains(t, renderedData, "value: \"250m\"")
-	
+
 	assert.Contains(t, renderedData, "name: deploy-to-production")
 	assert.Contains(t, renderedData, "value: production")
 	assert.Contains(t, renderedData, "value: prod-cluster")
 	assert.Contains(t, renderedData, "value: app-prod")
 	assert.Contains(t, renderedData, "value: \"3\"")
 	assert.Contains(t, renderedData, "value: \"1000m\"")
-	
+
 	// Verify features map iteration
 	assert.Contains(t, renderedData, "logging: true")
 	assert.Contains(t, renderedData, "monitoring: true")
 	assert.Contains(t, renderedData, "tracing: true")
-	
+
 	// Verify service config rendering in JSON format
 	assert.Contains(t, renderedData, `"name": "web-frontend"`)
 	assert.Contains(t, renderedData, `"port": 80`)
 	assert.Contains(t, renderedData, `"targetPort": 8080`)
 	assert.Contains(t, renderedData, `"type": "ClusterIP"`)
 	assert.Contains(t, renderedData, `"prometheus.io/scrape": "true"`)
-	
+
 	assert.Contains(t, renderedData, `"name": "api-backend"`)
 	assert.Contains(t, renderedData, `"port": 443`)
 	assert.Contains(t, renderedData, `"targetPort": 8443`)
@@ -807,12 +807,12 @@ spec:
 `,
 		},
 	}
-	
+
 	// Create resolver with mock fetcher
 	r := &resolver{
 		fetcher: mockData,
 	}
-	
+
 	// Test with complex object parameters
 	params := []pipelinev1.Param{
 		{
@@ -894,17 +894,17 @@ spec:
 			},
 		},
 	}
-	
+
 	// Execute the Resolve function
 	result, err := r.Resolve(context.Background(), params)
-	
+
 	// Verify results
 	require.NoError(t, err)
 	require.NotNil(t, result)
-	
+
 	// Check that the YAML was correctly rendered without property enumeration
 	renderedData := string(result.Data())
-	
+
 	// Check for direct rendering of validation steps
 	assert.Contains(t, renderedData, "name: security-validation")
 	assert.Contains(t, renderedData, "runAfter:")
@@ -913,7 +913,7 @@ spec:
 	assert.Contains(t, renderedData, "- pci-dss")
 	assert.Contains(t, renderedData, "- hipaa")
 	assert.Contains(t, renderedData, "description: Whether the deployment is compliant")
-	
+
 	// Check for rendering of resource configurations
 	assert.Contains(t, renderedData, "type: compute")
 	assert.Contains(t, renderedData, "name: app-server")
